Front: New Controls: Rename files, shorten paths
