+++ front/src/components/infinite/resource/album.tsx
