+++ front/src/components/infinite/resource/song.tsx
