+++ front/src/components/infinite/resource/video.tsx
