+++ front/src/pages/artists/[slugOrId]/videos.tsx
@@ -26,7 +26,7 @@ import {
 	getOrderQuery,
 	getSortQuery,
 } from "../../../components/infinite/controls/sort";
-import InfiniteVideoView from "../../../components/infinite/infinite-resource-view/infinite-video-view";
+import InfiniteVideoView from "../../../components/infinite/resource/video";
 import ArtistRelationPageHeader from "../../../components/relation-page-header/artist-relation-page-header";
 import { VideoSortingKeys } from "../../../models/video";
 import type { GetPropsTypesFrom, Page } from "../../../ssr";
