+++ front/src/pages/genres/[slugOrId].tsx
@@ -23,9 +23,9 @@ import { useTranslation } from "react-i18next";
 import API from "../../api/api";
 import { useQuery } from "../../api/use-query";
 import { Head } from "../../components/head";
-import InfiniteAlbumView from "../../components/infinite/infinite-resource-view/infinite-album-view";
-import InfiniteArtistView from "../../components/infinite/infinite-resource-view/infinite-artist-view";
-import { InfiniteSongView } from "../../components/infinite/infinite-resource-view/infinite-song-view";
+import InfiniteAlbumView from "../../components/infinite/resource/album";
+import InfiniteArtistView from "../../components/infinite/resource/artist";
+import { InfiniteSongView } from "../../components/infinite/resource/song";
 import { useTabRouter } from "../../components/tab-router";
 import type { GetPropsTypesFrom, Page } from "../../ssr";
 import getSlugOrId from "../../utils/getSlugOrId";
