+++ front/src/pages/songs/[slugOrId]/[...tab].tsx
@@ -41,9 +41,9 @@ import ExternalMetadataBadge from "../../../components/external-metadata-badge";
 import GenreButton from "../../../components/genre-button";
 import { Head } from "../../../components/head";
 import { PlayIcon } from "../../../components/icons";
-import { InfiniteSongView } from "../../../components/infinite/infinite-resource-view/infinite-song-view";
-import InfiniteTrackView from "../../../components/infinite/infinite-resource-view/infinite-track-view";
-import InfiniteVideoView from "../../../components/infinite/infinite-resource-view/infinite-video-view";
+import { InfiniteSongView } from "../../../components/infinite/resource/song";
+import InfiniteTrackView from "../../../components/infinite/resource/track";
+import InfiniteVideoView from "../../../components/infinite/resource/video";
 import LyricsBox from "../../../components/lyrics";
 import SongRelationPageHeader from "../../../components/relation-page-header/song-relation-page-header";
 import SongTypeIcon from "../../../components/song-type-icon";
