+++ front/src/pages/songs/index.tsx
@@ -6,7 +6,7 @@ import {
 	getOrderQuery,
 	getSortQuery,
 } from "../../components/infinite/controls/sort";
-import { HybridInfiniteSongView } from "../../components/infinite/infinite-resource-view/infinite-song-view";
+import { HybridInfiniteSongView } from "../../components/infinite/resource/song";
 import { SongSortingKeys } from "../../models/song";
 import type { GetPropsTypesFrom, Page } from "../../ssr";
